requirements.txt
pipecat-ai[webrtc,daily,openai,deepgram,cartesia,silero,tracing]
uvicorn
fastapi
python-dotenv
pipecatcloud
pipecat_ai_small_webrtc_prebuilt
pyaudio
opentelemetry-exporter-otlp-proto-http


